+++ src/_cffi_src/openssl/x509v3.py
@@ -190,18 +190,15 @@ void *X509V3_EXT_d2i(X509_EXTENSION *);
 MACROS = """
 /* This is a macro defined by a call to DECLARE_ASN1_FUNCTIONS in the
    x509v3.h header. */
-int i2d_BASIC_CONSTRAINTS(BASIC_CONSTRAINTS *, unsigned char **);
 BASIC_CONSTRAINTS *BASIC_CONSTRAINTS_new(void);
 void BASIC_CONSTRAINTS_free(BASIC_CONSTRAINTS *);
 /* This is a macro defined by a call to DECLARE_ASN1_FUNCTIONS in the
    x509v3.h header. */
 AUTHORITY_KEYID *AUTHORITY_KEYID_new(void);
 void AUTHORITY_KEYID_free(AUTHORITY_KEYID *);
-int i2d_AUTHORITY_KEYID(AUTHORITY_KEYID *, unsigned char **);
 
 NAME_CONSTRAINTS *NAME_CONSTRAINTS_new(void);
 void NAME_CONSTRAINTS_free(NAME_CONSTRAINTS *);
-int Cryptography_i2d_NAME_CONSTRAINTS(NAME_CONSTRAINTS *, unsigned char **);
 
 OTHERNAME *OTHERNAME_new(void);
 void OTHERNAME_free(OTHERNAME *);
@@ -214,11 +211,6 @@ void *X509V3_set_ctx_nodb(X509V3_CTX *);
 int i2d_GENERAL_NAMES(GENERAL_NAMES *, unsigned char **);
 GENERAL_NAMES *d2i_GENERAL_NAMES(GENERAL_NAMES **, const unsigned char **,
                                  long);
-
-int i2d_EXTENDED_KEY_USAGE(EXTENDED_KEY_USAGE *, unsigned char **);
-
-int i2d_AUTHORITY_INFO_ACCESS(Cryptography_STACK_OF_ACCESS_DESCRIPTION *,
-                              unsigned char **);
 
 int sk_GENERAL_NAME_num(struct stack_st_GENERAL_NAME *);
 int sk_GENERAL_NAME_push(struct stack_st_GENERAL_NAME *, GENERAL_NAME *);
@@ -267,9 +259,6 @@ void NOTICEREF_free(NOTICEREF *);
 USERNOTICE *USERNOTICE_new(void);
 void USERNOTICE_free(USERNOTICE *);
 
-int i2d_CERTIFICATEPOLICIES(Cryptography_STACK_OF_POLICYINFO *,
-                            unsigned char **);
-
 void sk_POLICYQUALINFO_free(Cryptography_STACK_OF_POLICYQUALINFO *);
 int sk_POLICYQUALINFO_num(Cryptography_STACK_OF_POLICYQUALINFO *);
 POLICYQUALINFO *sk_POLICYQUALINFO_value(Cryptography_STACK_OF_POLICYQUALINFO *,
@@ -303,17 +292,7 @@ void DIST_POINT_free(DIST_POINT *);
 DIST_POINT_NAME *DIST_POINT_NAME_new(void);
 void DIST_POINT_NAME_free(DIST_POINT_NAME *);
 
-int i2d_CRL_DIST_POINTS(Cryptography_STACK_OF_DIST_POINT *, unsigned char **);
 """
 
 CUSTOMIZATIONS = """
-/* i2d_NAME_CONSTRAINTS doesn't exist, but this is the way the macros in
-   asn1t.h would implement it. We're not using those macros in case
-   OpenSSL exposes this function in the future. */
-int Cryptography_i2d_NAME_CONSTRAINTS(NAME_CONSTRAINTS *nc,
-                                      unsigned char **out) {
-    return ASN1_item_i2d((ASN1_VALUE *)nc, out,
-                         ASN1_ITEM_rptr(NAME_CONSTRAINTS));
-}
-
 """