+++ src/cryptography/utils.py
@@ -12,9 +12,11 @@ import sys
 import warnings
 
 
-# the functions deprecated in 1.0 are on an arbitrarily extended deprecation
-# cycle and should not be removed until we agree on when that cycle ends.
+# the functions deprecated in 1.0 and 1.4 are on an arbitrarily extended
+# deprecation cycle and should not be removed until we agree on when that cycle
+# ends.
 DeprecatedIn10 = DeprecationWarning
+DeprecatedIn14 = DeprecationWarning
 
 
 def read_only_property(name):